[WIP] Start to build TCP connection and packet framing
This patch starts to build TCP sessions and packet framing into ExWire. Once we get done framing (and protocol negotitation), we'll have a more or less working RLPx implementation (though, there's lot of details to fill in). Currently, a lot of code is dumped where it needs to be to get things to work (e.g. TCP should really call our to stub code). Additionally, we're only just beginning to get the framing code to work (the spec is, yet again, under-documented or strictly incorrect, so we're basing it off a mix of Geth and Parity's code). Luckily, geth included a test case so that's good. All in all, at this point, we can connect to a TCP node, negotiate keys and try to send packets. The key exchange (ECIES) looks to be right based on test cases, but I'm not 100% positive the derived keys are correct yet (see the TCP module). Luckily, the clients are pretty happy to accept and report any isuses, so we'll be able to diagnose anything that's incorrect after we finish our framing and de-framing code.